NLP_DEMO
This paper describes the model of text emotional classification. The data set is evaluated by JD commodity. The model uses Jieba participle to construct feature vectors. The classifier uses linear SVC to classify the data, and the classification accuracy reaches 88%. Source code includes code to crawl JD commodities, but IP address can only be crawled once, and score = 1 in URL links is a bad evaluation, score = 3 is a high-quality evaluation.